fixed rebuildtags.sh

+++ bin/rebuildtags.sh
@@ -1,8 +1,20 @@
 #!/bin/bash
 cd ../
-rm cscope*
-rm tags
-find ./ -name "*.go" > ./cscope.files
-cscope -Rbqk
-ctags -R
+path=`pwd`
+echo "rebuild path:"$path
+files=$(ls $path)
+for filename in $files
+do
+	if [[ ${filename:0:6} -eq "cscope" ]];then
+		rm cscope*
+		break
+	fi
+done
 
+if [ -f tags ];then
+	rm tags
+fi
+find ./ -name "*.go" > ./cscope.files
+cscope -Rbq
+ctags -R
+echo "rebuild succeed!"
